+++ rust_tracer/src/math.rs
@@ -28,7 +28,7 @@ pub fn sample_uniform_disk_concentric(u: (f32, f32)) -> (f32, f32) {
 pub fn sample_cosine_hemisphere(u: (f32, f32)) -> Vec3 {
     let d = sample_uniform_disk_concentric(u);
     let z = (1.0 - d.0 * d.0 - d.1 * d.1).sqrt().max(0.0);
-    Vec3::new(d.0 as f64, d.1 as f64, z as f64)
+    Vec3::new(d.0, d.1, z)
 }
 
 pub fn cosine_hemisphere_pdf(cos_theta: f32) -> f32 {
